fixed scripts

+++ deployment_scripts/gpu-setup.sh
@@ -2,20 +2,29 @@
 
 set -euo pipefail
 
+MARKER_FILE="/tmp/nvidia_after_reboot_done"
+
+sudo apt update
+sudo apt install -y pciutils 
+
 echo "Checking for NVIDIA GPU on this machine..."
 if ! lspci | grep -i nvidia; then
   echo "No NVIDIA GPU detected. Skipping driver + toolkit installation."
   exit 0
 fi
 
-sudo apt update
-
 echo "Installing NVIDIA drivers"
 # here since nvidia-driver-535 may not exist in all ubuntu versions so I'm using autoinstall
-sudo apt-get install -y ubuntu-drivers-common
-sudo ubuntu-drivers autoinstall
+if [ ! -f "$MARKER_FILE" ]; then
+  echo "Installing NVIDIA drivers..."
+  sudo apt-get install -y ubuntu-drivers-common
+  sudo ubuntu-drivers autoinstall
 
-#after installation of any drivers there is system reboot needed so do a reboot after completing last step
+  echo "Rebooting system to apply drivers..."
+  sudo touch "$MARKER_FILE"
+  sudo reboot
+  exit 0
+fi
 
 echo "Installing NVIDIA Container Toolkit" # it is importent for running docker containers.
